Code with analysis of parking incidents in 10019
In a second Jupyter notebook called Parking.ipynb, analyze whether illegal parking incidents are a larger fraction of total 311 incidents in your ZIP code than they are in general. Specifically, compute the total number of parking incidents in your ZIP, and the total number of all incidents -- is this fraction greater than or smaller than the total fraction of parking incidents across all ZIP codes? Your answer should be in the form of a single bool called `higher_parking_proportion` which is True if your ZIP contains a higher proportion of parking incidents than the global value, and False otherwise. For instance, if there were 200 parking incidents in your ZIP, and 1000 total incidents in your ZIP, a rate of 20%, but the global parking incident rate was 3000 parking incidents out of 10000 total, a ratio of 30%, you would assign higher_parking_proportion = False.